+++ migrations/M1624120320_add_booster_tables.py
@@ -1,0 +1,10 @@
+from model.model import *
+
+factory = DatabaseFactory()
+connection = factory.get_database_connection()
+
+connection.create_tables([
+    BoosterMessage,
+    BoosterRoleConfig,
+    BoosterRole,
+])